% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_espn_projections.R
\name{read.espn.projections}
\alias{read.espn.projections}
\title{Reads projections from ESPN.}
\usage{
read.espn.projections()
}
\description{
Reads projections from ESPN.
}
\examples{
read.espn.projections()
}

